A very early version for upload
currently can compile but can not insert data to mysql (connection is ok though), might be a string overflow problem, can be fixed.

Next step is to make sure this script is usable and can work with hardware part by provide suitable data insert interface in prepare part of code.